+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `passwordMode` property to connection schema. It can be set to use
+  `sqltools-driver-credentials` authentication provider.
+
 ### Changed
 
-- Updated deps.
+- Refactor TLS options in connection schema.
+- Updated dependencies.
 
 ## [0.4.2] - 2023-08-31